🎨 Session components naming update
