+++ packages/core/src/domain/session/sessionCookieStore.ts
@@ -3,7 +3,7 @@ import { COOKIE_ACCESS_DELAY, deleteCookie, getCookie, setCookie } from '../../b
 import { isChromium } from '../../tools/utils/browserDetection'
 import { objectEntries } from '../../tools/utils/polyfills'
 import { SESSION_EXPIRATION_DELAY } from './sessionConstants'
-import type { SessionState, SessionStorage } from './sessionStorage'
+import type { SessionState, SessionStore } from './sessionStore'
 
 const SESSION_ENTRY_REGEXP = /^([a-z]+)=([a-z0-9-]+)$/
 const SESSION_ENTRY_SEPARATOR = '&'
@@ -14,9 +14,9 @@ export const SESSION_COOKIE_NAME = '_dd_s'
 export const LOCK_RETRY_DELAY = 10
 export const MAX_NUMBER_OF_LOCK_RETRIES = 100
 
-export function initCookieStorage(options: CookieOptions): SessionStorage {
+export function initCookieStore(options: CookieOptions): SessionStore {
   return {
-    storageAccessOptions: {
+    storeAccessOptions: {
       pollDelay: COOKIE_ACCESS_DELAY,
       /**
        * Cookie lock strategy allows mitigating issues due to concurrent access to cookie.
